Put divs into flexbox and grid
The flexbox make the exercise responsive to smaller screens (sort of,
still pretty ugly).  The grid made it so I could start to line up
some of the UI elements better.

But I think I'm going to drop the flexbox for a fully gridified layout
next, to further align UI bits with each other.